Clean up. Create history layout. Create history route to route all user history related routes to single routes since they share most of their functions and ui.
