+++ prisma/schema.prisma
@@ -1,6 +1,3 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
 generator client {
   provider        = "prisma-client-js"
   previewFeatures = ["referentialIntegrity"]
@@ -8,7 +5,7 @@ generator client {
 
 datasource db {
   provider             = "mysql"
-  url                  = env("DEV_DATABASE_URL")
+  url                  = env("DATABASE_URL")
   referentialIntegrity = "prisma"
 }
 
@@ -25,8 +22,7 @@ model Account {
   scope             String?
   id_token          String? @db.Text
   session_state     String?
-
-  user User @relation(fields: [userId], references: [id], onDelete: Cascade)
+  user              User    @relation(fields: [userId], references: [id], onDelete: Cascade)
 
   @@unique([provider, providerAccountId])
 }
@@ -107,26 +103,6 @@ model Comment {
   Post             Post?              @relation(fields: [postId], references: [id])
   postId           String?
   userCommentSaved UserCommentSaved[]
-
-  mainCommentId String?
-
-  replyId String?
-
-  // @@unique([id, author])
+  mainCommentId    String?
+  replyId          String?
 }
-
-// model Reply {
-//   id            String    @id
-//   body          String    @db.LongText
-//   bodyHtml      String    @db.LongText
-//   author        String
-//   score         Int
-//   created       DateTime
-//   updatedAt     DateTime? @updatedAt
-//   parentComment Comment?    @relation(fields: [commentId], references: [id])
-//   commentId       String?
-
-//   replies Reply[] @relation("replies")
-//   Reply   Reply?  @relation(name: "replies", fields: [replyId], references: [id], onDelete: NoAction, onUpdate: NoAction)
-//   replyId String?
-// }
