+++ src/pages/api/auth/[...nextauth].ts
@@ -76,6 +76,7 @@ export default async function auth(req: NextApiRequest, res: NextApiResponse) {
                 return { ...session, user: user };
             },
             signIn: async ({ user }) => {
+                console.log("Sign in callback")
                 if (user.id) {
                     const validEmail = await prisma.user.findUnique({
                         where: {
